Add heap route
