+++ back/src/logging/logger.ts
@@ -1,9 +1,5 @@
 import appRoot from "app-root-path";
 import { createLogger, format, transports } from "winston";
-import v8 from "v8";
-import { createWriteStream } from "fs";
-import { S3Client } from "@aws-sdk/client-s3";
-import { Upload } from "@aws-sdk/lib-storage";
 
 const LOG_PATH = `${appRoot}/logs/app.log`;
 // Avoid using undefined console.log() in jest context
@@ -34,49 +30,5 @@ const logger = createLogger({
       : new transports.File({ filename: LOG_PATH })
   ]
 });
-
-createHeapSnapshotAndUploadToS3();
-
-async function createHeapSnapshotAndUploadToS3() {
-  if (process.env.DEBUG_HEAPDUMP !== "active") {
-    return;
-  }
-  const snapshotStream = v8.getHeapSnapshot();
-  // It's important that the filename end with `.heapsnapshot`,
-  // otherwise Chrome DevTools won't open it.
-  const fileName = `${process.env.DD_ENV}_${
-    process.env.CONTAINER
-  }_${Date.now()}.heapsnapshot`;
-  const fileStream = createWriteStream(fileName);
-  snapshotStream.pipe(fileStream);
-
-  try {
-    const parallelUploads3 = new Upload({
-      client: new S3Client({
-        endpoint: process.env.S3_ENDPOINT,
-        region: process.env.S3_REGION,
-        credentials: {
-          accessKeyId: process.env.S3_ACCESS_KEY_ID,
-          secretAccessKey: process.env.S3_SECRET_ACCESS_KEY
-        }
-      }),
-      params: {
-        Bucket: process.env.S3_BUCKET,
-        Key: fileName,
-        Body: snapshotStream
-      },
-      leavePartsOnError: false
-    });
-
-    parallelUploads3.on("httpUploadProgress", progress => {
-      logger.info(progress);
-    });
-
-    await parallelUploads3.done();
-  } catch (e) {
-    logger.error("Error while uploading heapdump", e);
-  }
-  setTimeout(() => createHeapSnapshotAndUploadToS3(), 1000 * 60 * 60);
-}
 
 export default logger;
